+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Added Configuration option to hide "Drop Item" message in chat.
+- Changed many effect descriptions to be more descriptive.
 
 ## [1.0.5] - 2022-07-30